fix bug

+++ WriteALogPackage/internal/log/store.go
@@ -39,7 +39,7 @@ func (s *store) Append(p []byte) (n uint64, pos uint64, err error) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 	pos = s.size
-	if err := binary.Write(s.buf, enc, uint(len(p))); err != nil {
+	if err := binary.Write(s.buf, enc, uint64(len(p))); err != nil {
 		return 0, 0, err
 	}
 	w, err := s.buf.Write(p)
